refactor: Move reactions module inside posts folder
