+++ src/modules/posts/reactions/types/request.ts
@@ -1,0 +1,4 @@
+export type ReactPostParams = {
+	postId: number;
+	userId: number;
+};
