+++ src/modules/posts/service.ts
@@ -10,11 +10,7 @@ import {
 	CouldNotDeleteError,
 	CouldNotUpdateError,
 } from "@shared/types/errors";
-import type {
-	CreatePostParams,
-	ReactPostParams,
-	UpdatePostParams,
-} from "./types/request";
+import type { CreatePostParams, UpdatePostParams } from "./types/request";
 
 export const getPostById = async (id: number) => {
 	const result = await db
@@ -96,44 +92,6 @@ export const createPost = async (post: CreatePostParams) => {
 	if (!result.length) {
 		throw new CouldNotCreateError(
 			`Failed to create post with content: ${post.content} for user: ${post.userId}`,
-		);
-	}
-
-	return result[0];
-};
-
-export const reactPost = async (reaction: ReactPostParams) => {
-	const result = await db
-		.insert(Reactions)
-		.values({
-			userId: reaction.userId,
-			postId: reaction.postId,
-		})
-		.returning();
-
-	if (!result.length) {
-		throw new CouldNotCreateError(
-			`Failed to create reaction for post with id: ${reaction.postId} and user with id: ${reaction.userId}`,
-		);
-	}
-
-	return result[0];
-};
-
-export const unreactPost = async (reaction: ReactPostParams) => {
-	const result = await db
-		.delete(Reactions)
-		.where(
-			and(
-				eq(Reactions.userId, reaction.userId),
-				eq(Reactions.postId, reaction.postId),
-			),
-		)
-		.returning();
-
-	if (!result.length) {
-		throw new CouldNotDeleteError(
-			`Failed to delete reaction for post with id: ${reaction.postId} and user with id: ${reaction.userId}`,
 		);
 	}
